Create a JavaScript consumer
Both the steps that I had to perform and the code in my repository
**differ** from the video.

When I execute `yarn link` in `mylib`, yarn reports an error that the
package is already linked. I ignored the error, linked back to the
function `introduceMyself` in `mylib/src/index.ts` in the `js-consumer`
directory by executing `yarn link mylib` in that directory. This commend
succeeds and the link is apparently established. Perhaps a Mac / Windows
difference; perhaps a newer version of yarn.

The code as in my repository **differs** from the code on the video.
Specifically, the video indicates a need to have the `main` target in
the `package.json` file of `mylib` to have the value `./dist/main.js`. I
seem to not need this seem link with `yarn` 1.22.19. (I believe that
Jack was using `yarn` 1.22.10.)

The joys of the JavaScript / TypeScript environment?